Move all of BaseImporter content (all static methods) to a module
